requirements.txt
requests
jinja2
lxml
python-jenkins